Карты. Не дооформлен личный список карт. Отсутствует полный спиок всех карт всех пользователей.

+++ app/Http/Livewire/CreateNew/NewMap.php
@@ -13,14 +13,7 @@ class NewMap extends Component
 {
     use WithFileUploads;
 
-    public $description='';
-    public $title='';
-    public $map_data;
-    public $map;
-    public $post;
-    public $photo;
-    public $name;
-    public $action;
+    public $map_data, $map, $post, $photo, $name, $action, $user, $description, $title;
 
     protected $messages = [
         'title.required' => 'Введите название карты.',
@@ -28,45 +21,49 @@ class NewMap extends Component
         'map_data.min' => 'Нарисуйте свой маршрут',
     ];
 
-    public function mount()
+    public function mount(User $user, $action, Map $map)
     {
-            $this->getMap();
+        abort_if($this->user->cant('update', $this->map), 403, 'Это не ваша карта!');
+        $this->action = $action;
+        $this->map = $map;
+        $this->setMapAttributes();
+        $this->user=$user;
     }
 
-    public function getMap()
+    protected function setMapAttributes()
     {
         if($this->action == 'create'){
-            $this->map = User::find(auth()->user()->id)->maps()->create();
-        }else{
-            $this->map = User::find(auth()->user()->id)->maps()->orderBy('created_at', 'DESC')->first();
+            $this->title = '';
+            $this->description = '';
+        } elseif($this->action == 'edit') {
             $this->title = $this->map->title;
             $this->description = $this->map->post->text;
+        } else{
+            abort(404,'Нет такой страницы');
         }
     }
 
     public function saveDraft()
     {
         $this->saveMap();
-        $this->emitUp('showFeed');
     }
 
     public function publishMap()
     {
         $this->saveMap();
         $this->map->update(['published' => true]);
-        return redirect()->route('map.index', [
-            'maps' => Map::orderBy('created_at', 'desc')->paginate(10)
-        ]);
+        return redirect()->route('user.maps', ['user' => auth()->user()->id]);
     }
 
     public function previewMap()
     {
         $this->saveMap();
-        $this->emit('showMap', $this->map->id);
+        return redirect()->route('user.map.show', ['user' => auth()->user()->id, 'mode' => 'preview', 'map' => $this->map]);
     }
 
     protected function saveMap()
     {
+        abort_if($this->user->cant('update', $this->map), 403, 'Вы не можете редактировать эту карту!');
         $this->validate([
             'title' => 'required|min:5',
             'map_data' => 'json|min:8',
@@ -78,8 +75,9 @@ class NewMap extends Component
 
     public function deleteMap()
     {
+        abort_if($this->user->cant('delete', $this->map), 403, 'Вы не имеете права удалять эту карту!');
         $this->map->delete();
-        $this->emit('showFeed');
+        return redirect()->route('user.maps', ['user' => auth()->user()->id]);
     }
 
     public function savePhoto()
@@ -106,6 +104,8 @@ class NewMap extends Component
 
     public function render()
     {
-        return view('livewire.create-new.new-map');
+        return view('livewire.create-new.new-map')
+            ->extends('layouts.app', ['user' => $this->user])
+            ->section('content');
     }
 }
